Migrating NETUG to App Router
